Display messages directly from Core Data
